Add validate logic and null check.

+++ taxi-service/taxi-service/src/main/java/intelligent_taxi/taxiservice/command/TaxiCommandService.java
@@ -1,11 +1,12 @@
 package intelligent_taxi.taxiservice.command;
 
+import intelligent_taxi.taxiservice.controller.restResponse.ResponseMessage;
 import intelligent_taxi.taxiservice.domain.Taxi;
 import intelligent_taxi.taxiservice.dto.TaxiRequest;
 import intelligent_taxi.taxiservice.dto.UpdateGradeRequest;
 import intelligent_taxi.taxiservice.dto.UpdateRegionRequest;
+import intelligent_taxi.taxiservice.exception.TaxiCustomException;
 import intelligent_taxi.taxiservice.repository.TaxiRepository;
-import intelligent_taxi.taxiservice.validator.ServiceValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,7 +17,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class TaxiCommandService {
 
     private final TaxiRepository taxiRepository;
-    private final ServiceValidator serviceValidator;
 
     public Long createTaxi(TaxiRequest requestDto, String username) {
         Taxi taxi = Taxi.create(requestDto, username);
@@ -24,22 +24,31 @@ public class TaxiCommandService {
     }
 
     public void updateRegion(UpdateRegionRequest requestDto, String username, Long id) {
-        Taxi taxi = serviceValidator.validateTaxiAndUsername(username, id);
-        taxi.updateRegion(requestDto.getRegion());
+        Taxi taxi = taxiRepository.findOneById(id)
+                .orElseThrow(() -> new TaxiCustomException(ResponseMessage.TAXI_IS_NULL));
+        taxi.updateRegion(requestDto.getRegion(), username);
     }
 
     public void updateGrade(UpdateGradeRequest requestDto, String username, Long id) {
-        Taxi taxi = serviceValidator.validateTaxiAndUsername(username, id);
-        taxi.updateGrade(requestDto.getTaxiGrade());
+        Taxi taxi = taxiRepository.findOneById(id)
+                .orElseThrow(() -> new TaxiCustomException(ResponseMessage.TAXI_IS_NULL));
+        taxi.updateGrade(requestDto.getTaxiGrade(), username);
     }
 
     public void deleteOneByUsername(String username) {
-        Taxi taxi = serviceValidator.validateDeleteOneByUsername(username);
+        Taxi taxi = taxiRepository.findOneByUsername(username)
+                .orElseThrow(() -> new TaxiCustomException(ResponseMessage.TAXI_IS_NULL));
         taxiRepository.delete(taxi);
     }
 
     public void deleteOneById(Long id, String username) {
-        Taxi taxi = serviceValidator.validateTaxiAndUsername(username, id);
+        Taxi taxi = taxiRepository.findOneById(id)
+                .orElseThrow(() -> new TaxiCustomException(ResponseMessage.TAXI_IS_NULL));
+
+        if (!taxi.getUsername().equals(username)) {
+            throw new TaxiCustomException(ResponseMessage.NOT_MATCH_USERNAME);
+        }
+
         taxiRepository.delete(taxi);
     }
 }
